requirements2.txt
absl-py==2.3.1
aiofiles==23.2.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
executing==2.2.1
fastapi==0.121.2
ffmpy==1.0.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
gast==0.6.0
git-filter-repo==2.47.0
google-pasta==0.2.0
gradio==6.3.0
gradio_client==2.0.3
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.11
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.1.0
narwhals==2.12.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==1.26.4
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.4
packaging==25.0
pandas==2.2.3
parso==0.8.5
pillow==10.4.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.0.2
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
safehttpx==0.1.7
scikit-fuzzy==0.5.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-cpu==2.17.1
tensorflow-intel==2.17.1
termcolor==3.2.0
threadpoolctl==3.6.0
tomlkit==0.13.3
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.21.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.8.9
uvicorn==0.38.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
